bot/requirements.txt
python-telegram-bot==20.8
httpx==0.26.0
python-dotenv==1.0.1
apscheduler==3.10.4
